scraper/requirements.txt
playwright==1.47.0
httpx==0.27.2
supabase==2.7.4
python-dotenv==1.0.1
